+++ internal/adapters/encryption/plain_builder/builder.go
@@ -39,6 +39,14 @@ func (b *plainBuilder) SetDatabasePart(ctx context.Context, identifier string) e
 	return nil
 }
 
+func (b *plainBuilder) GetProjectPart(_ context.Context) string {
+	return b.projectPart
+}
+
+func (b *plainBuilder) GetDatabasePart(_ context.Context) string {
+	return b.databasePart
+}
+
 func (b *plainBuilder) Build(_ context.Context) (string, error) {
 	if b.projectPart == "" {
 		return "", domainErrors.ErrProjectPartRequired