+++ internal/core/ports/builders/encryption.go
@@ -7,5 +7,9 @@ import (
 type EncryptionKeyBuilder interface {
 	SetProjectPart(ctx context.Context, identifier string) error
 	SetDatabasePart(ctx context.Context, identifier string) error
+
+	GetProjectPart(ctx context.Context) string
+	GetDatabasePart(ctx context.Context) string
+
 	Build(ctx context.Context) (string, error)
 }