+++ pkg/cypher/cypher.go
@@ -103,9 +103,9 @@ func ReconstructEncryptionKey(part1, part2 string) (string, error) {
 		return "", err
 	}
 
-	subset := make(map[byte][]byte)
-	subset[1] = rawPart1
-	subset[2] = rawPart2
+	subset := make(map[byte][]byte, 2)
+	subset[0] = rawPart1
+	subset[1] = rawPart2
 
 	key := sss.Combine(subset)